Catalyst needs to be loaded for us to load.
Our monkey-patch pokes at Catalyst, so Catalyst needs to be loaded
first.

In the usual way of loading the plugin, e.g. listing the plugins you
want on the `use Catalyst` line, that's fine, but here we're testing
just that the plugin compiles, so we will need to load Catalyst first.